separate subdags as jinja template include
That allows to segregate without having to alter the api controller
logic.
