+++ breeder/linux_network_stack/root_dag.py
@@ -73,271 +73,13 @@ DEFAULTS = {
     # 'sla_miss_callback': yet_another_function,
     }
 
-## coroutines
 NATS_SERVER_URL = "nats://godon_nats_1:4222"
-# interaction
-async def gather_instruction():
-    # Connect to NATS Server.
-    nc = await nats.connect(NATS_SERVER_URL)
-    sub = await nc.subscribe('effectuation')
-    msg = await sub.next_msg(timeout=300)
-    print(msg)
-    await msg.respond(b'OK')
-    await nc.close()
-    return msg.data.decode()
-
-{% raw %}
-async def deliver_probe(metric_value):
-    # Connect to NATS Server.
-    nc = await nats.connect(NATS_SERVER_URL)
-    metric_data = dict(metric=metric_value)
-    transmit_data = bytes(json.dumps(metric_data), encoding='utf8')
-    while True:
-        try:
-            response = await nc.request('recon', transmit_data)
-            print('Response:', response )
-            break
-        except nats.errors.NoRespondersError:
-            time.sleep(2)
-            continue
-        except:
-            logger.warning('unexpted exception')
-            logger.warning(sys.exc_info()[0])
-            raise
-    await nc.flush()
-    await nc.close()
-{% endraw %}
-# optimization
-
-
-def create_target_interaction_dag(dag_id, config):
-
-    dag = DAG(dag_id,
-              default_args=DEFAULTS,
-              description='breeder subdag for interacting with targets')
-
-    with dag as interaction_dag:
-
-        dump_config = BashOperator(
-            task_id='print_config',
-            bash_command='echo ${config}',
-            env={"config": str(config)},
-            dag=interaction_dag,
-        )
-
-        @dag.task(task_id="pull_optimization_step")
-        def run_pull_optimization():
-            task_logger.debug("Entering")
-
-            msg = asyncio.run(gather_instruction())
-            settings = json.loads(msg).get('settings')
-
-            task_logger.debug(f"Settings: f{settings}")
-
-            return settings
-
-        pull_step = run_pull_optimization()
-
-        @dag.task(task_id="push_optimization_step")
-        def run_push_optimization(ti=None):
-            task_logger.debug("Entering")
-
-            metric_value = ti.xcom_pull(task_ids="recon_step")
-
-            task_logger.debug(f"Metric : f{metric_value}")
-
-            msg = asyncio.run(deliver_probe(metric_value))
-
-            task_logger.debug("Done")
-
-            return msg
-
-        push_step = run_push_optimization()
-
-        @dag.task(task_id="recon_step")
-        def run_reconnaissance():
-            task_logger.debug("Entering")
-            prom_conn = PrometheusConnect(url ="http://godon_prometheus_1:9090", disable_ssl=True)
-
-            start_time = parse_datetime("2m")
-            end_time = parse_datetime("now")
-            chunk_size = timedelta(minutes=1)
-
-            metric_data = dict()
-            for query in config.get('recon').get('prometheus'):
-                query_name = query.get('name')
-                query_string = query.get('query')
-                query_result = prom_conn.custom_query(query_string)
-                metric_value = query_result[0]
-                metric_data[query_name] = metric_value.get('value')[1]
-
-            task_logger.debug("Done")
-
-            return metric_data
-
-        recon_step = run_reconnaissance()
-
-        _ssh_hook = SSHHook(
-            remote_host=config.get('effectuation').get('target'),
-            username=config.get('effectuation').get('user'),
-            key_file=config.get('effectuation').get('key_file'),
-            timeout=30,
-            keepalive_interval=10
-        )
-
-{% raw %}
-        effectuation_step = SSHOperator(
-            ssh_hook=_ssh_hook,
-            task_id='effectuation',
-            timeout=30,
-            command="""
-                    {{ ti.xcom_pull(task_ids='pull_optimization_step') }}
-                    """,
-            dag=interaction_dag,
-        )
-{% endraw %}
-
-        @dag.task(task_id="run_iter_count_step")
-        def run_iter_count(ti=None):
-            last_iteration =  ti.xcom_pull(task_ids="run_iter_count_step")
-            current_iteration = last_iteration + 1 if last_iteration else 0
-            return current_iteration
-
-        run_iter_count_step = run_iter_count()
-
-        @task.branch(task_id="stopping_decision_step")
-        def stopping_decision(max_iterations, ti=None):
-            task_logger.debug("Entering")
-            current_iteration = ti.xcom_pull(task_ids="run_iter_count_step")
-            def is_stop_criteria_reached(iteration):
-                if iteration >= max_iterations:
-                    return True
-                else:
-                    return False
-
-            task_logger.debug("Done")
-            if is_stop_criteria_reached(current_iteration):
-                return "stop_step"
-            else:
-                return "continue_step"
-
-        stopping_conditional_step = stopping_decision(config.get('run').get('iterations').get('max'))
-
-        continue_step = TriggerDagRunOperator(
-                task_id='continue_step',
-                trigger_dag_id=interaction_dag.dag_id,
-                dag=interaction_dag
-                )
-
-        stop_step = EmptyOperator(task_id="stop_task", dag=interaction_dag)
-
-        dump_config >> pull_step >> effectuation_step >> recon_step >> push_step >> run_iter_count_step >> stopping_conditional_step >> [continue_step, stop_step]
-
-    return dag
-
-
-def create_optimization_dag(dag_id, config):
-
-    dag = DAG(dag_id,
-              default_args=DEFAULTS,
-              description='breeder subdag for optimizing \
-                    linux network stack dynamics')
-
-    with dag as optimization_dag:
-
-        dump_config = BashOperator(
-            task_id='print_config',
-            bash_command='echo ${config}',
-            env={"config": str(config)},
-            dag=optimization_dag,
-        )
-
-        ## perform optimiziation run
-        @dag.task(task_id="optimization_step")
-        def run_optimization():
-
-            def objective(trial):
-                import logging
-                logger = logging.getLogger('objective')
-                logger.setLevel(logging.DEBUG)
-{% raw %}
-                async def do_effectuation(settings):
-                    import time
-                    import nats
-                    import sys
-                    # Connect to NATS Server.
-                    nc = await nats.connect(NATS_SERVER_URL)
-                    settings_data = dict(settings=settings)
-                    transmit_data = bytes(json.dumps(settings_data), encoding='utf8')
-                    while True:
-                        try:
-                            response = await nc.request('effectuation', transmit_data)
-                            print('Response:', response )
-                            break
-                        except nats.errors.NoRespondersError:
-                            time.sleep(2)
-                            continue
-                        except:
-                            logger.warning('unexpted exception')
-                            logger.warning(sys.exc_info()[0])
-                            raise
-                    await nc.flush()
-                    await nc.close()
-{% endraw %}
-
-                async def gather_recon():
-                    # Connect to NATS Server.
-                    nc = await nats.connect(NATS_SERVER_URL)
-                    sub = await nc.subscribe('recon')
-                    msg = await sub.next_msg(timeout=300)
-                    print(msg)
-                    await msg.respond(b'OK')
-                    await nc.close()
-                    return msg.data.decode()
-
-                logger.warning('entering')
-
-                settings = []
-                for setting_name, setting_config in config.get('settings').get('sysctl').items():
-                    constraints = setting_config.get('constraints')
-                    suggested_value = trial.suggest_int(setting_name, constraints.get('lower') , constraints.get('upper') )
-                    if setting_name in ['net.ipv4.tcp_rmem', 'net.ipv4.tcp_wmem']:
-                        settings.append(f"sudo sysctl -w {setting_name}='4096 131072 {suggested_value}';")
-                    else:
-                        settings.append(f"sudo sysctl -w {setting_name}='{suggested_value}';")
-
-                settings = '\n'.join(settings)
-
-                logger.warning('doing effectuation')
-                asyncio.run(do_effectuation(settings))
-                logger.warning('gathering recon')
-                metric = json.loads(asyncio.run(gather_recon()))
-                metric_value = metric.get('metric')
-                rtt = float(metric_value['tcp_rtt'])
-                delivery_rate = float(metric_value['tcp_delivery_rate_bytes'])
-                logger.warning(f'metric received {metric_value}')
-                logger.warning('Done')
-
-                return rtt, delivery_rate
-
-            with Client(address="godon_dask_scheduler_1:8786") as client:
-                # Create a study using Dask-compatible storage
-                storage = DaskStorage(InMemoryStorage())
-                study = optuna.create_study(directions=["minimize", "maximize"], storage=storage)
-                # Optimize in parallel on your Dask cluster
-                futures = [
-                    client.submit(study.optimize, objective, n_trials=10, pure=False)
-                ]
-                wait(futures, timeout=7200)
-                print(f"Best params: {study.best_params}")
-
-        optimization_step = run_optimization()
-
-        dump_config >> optimization_step
-
-    return dag
 
 ###
+
+{% include 'effectuation.py' %}
+
+{% include 'optimization.py' %}
 
 config = {{ breeder }}
